pyproject.toml
[project]
name = "agents-crewai-tactical-multimodal"
version = "0.1.0"
description = "A multi-agent system for tactical analysis using CrewAI."
authors = [
    { name = "Pablo Martínez Agulló", email = "pablo.martinez.agullo@gmail.com" }
]
license = { text = "GNU General Public License (GPL) 3.0" }
requires-python = ">=3.11"
dependencies = [
    # Core CrewAI dependencies
    "crewai",
    "crewai[tools]",
    "openai",
    "python-dotenv",
    # Multimodal processing dependencies
    "openai-whisper", # Audio transcription
    "Pillow", # Image processing
    "pytesseract", # OCR text extraction
    "opencv-python", # Computer vision
    "PyPDF2", # PDF processing
    "numpy", # Numerical operations
    # Geolocation dependencies
    "geopy", # Geocoding and location services
    "requests", # HTTP requests for IP geolocation
    # UI dependencies
    "gradio",
    
    # MQTT
    "paho-mqtt>=2.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
]
